+++ .lintstagedrc.js
@@ -1,8 +1,7 @@
 module.exports = {
-  '**/*.css': ['stylelint'],
   '**/*.{js,ts,tsx}': [
     'eslint --cache --cache-location node_modules/.cache/eslint --cache-strategy content',
   ],
-  '**/*.{json,md,mdx,yml}': ['prettier --check'],
+  '**/*.{json,md,yml}': ['prettier --check'],
   '**/*.{ts,tsx}': () => 'pnpm lint:types',
 };