+++ packages/uniforms-unstyled/src/TextField.tsx
@@ -30,9 +30,7 @@ function Text({
         disabled={disabled}
         id={id}
         name={name}
-        onChange={event =>
-          onChange(event.target.value === '' ? undefined : event.target.value)
-        }
+        onChange={event => onChange(event.target.value)}
         placeholder={placeholder}
         readOnly={readOnly}
         ref={inputRef}